iio: adc: mcp320x: Support continuous conversion mode
The MCP3550/1/3 can be run in "continuous conversion mode", allowing
conversions to be fetched at any time without observing a delay,
at the expense of higher energy consumption.

The mode is enabled by permanently driving CS low, e.g. by wiring it to
ground.  This requires either exclusive bus usage by the chip or that
CLK and MISO are AND-gated with the CS signal coming from the master.

I only have boards available which use the chip in single conversion
mode, so cannot test its behaviour in continuous conversion mode.
I am putting this commit out there as a basis for others to work on.
The commit introduces a device property to indicate that continuous
conversion mode is enabled permanently.

It may be useful to support entering the mode temporarily, e.g. as long
as the chardev is opened:  Setting the cs_change bit of transfer[1] in
struct mcp320x should enter the mode, clearing the bit should exit it.

Signed-off-by: Lukas Wunner <lukas@wunner.de>