Bluetooth: hci_bcm: Handle intel-lpss based UART's too.
When the UART is part of an intel-lpss based mfd, the resulting device
tree has an extra level. Specifically we have the following layout (with
the /sys/devices/ prefix removed for brevity):

  ACPI device node:
    LNXSYSTM:00/LNXSYBUS:00/PNP0A08:00/device:94/BCM2E7C:00
  Associated physical device node:
    LNXSYSTM:00/LNXSYBUS:00/PNP0A08:00/device:94/BCM2E7C:00/physical_node -> pci0000:00/0000:00:1e.0/BCM2E7C:00/
  UART device node:
    pci0000:00/0000:00:1e.0/dw-apb-uart.2
  UART TTY device node:
    pci0000:00/0000:00:1e.0/dw-apb-uart.2/tty

This driver was assuming that physical device's parent is the UART, but
in this case the parent is the mfd device. We therefore check for both
cases.

Note that because this assumption appears in some user-space tools too,
it might be better to put the physical device node under the UART
instead of the mfd.

Signed-off-by: Ronald Tschalär <ronald@innovation.ch>