Bluetooth: hci_ldisc: Allow sleeping while proto locks are held.
Commit dec2c92880cc5435381d50e3045ef018a762a917 introduced locks around
the proto functions, using rwlock's, which do not allow sleeping while
the locks are held. However, the proto functions in hci_bcm use mutexes
and hence need to be able to sleep.

Therefore this replaces the rwlock's with rw_semaphore's. Because the
writes are very rare compared to the reads the percpu variant is used.

Lastly, the locks in the tx_wakeup callback needed to be removed because
that is called from an IRQ context. But since it doesn't actually call
any proto functions, instead just queueing work, and the
HCI_UART_PROTO_READY flag is checked (again) in the worker, this doesn't
cause any problems.

Signed-off-by: Ronald Tschalär <ronald@innovation.ch>